fix: lint error & tsc error
