+++ src/common/assets/colorMap.ts
@@ -1,8 +1,8 @@
 const colorMap: Record<string, string> = {
-    Crimson: '#C30003',
-    Black: 'black',
-    White: '#F2F2F2',
-    Ivory: '#F2E8DA',
-}
+  Crimson: '#C30003',
+  Black: 'black',
+  White: '#F2F2F2',
+  Ivory: '#F2E8DA',
+};
 
 export default colorMap;
