+++ src/common/assets/iconMap.ts
@@ -1,18 +1,18 @@
 const iconMap: Record<string, string> = {
-    축구: 'mgc_football_fill',
-    농구: 'mgc_basketball_fill',
-    야구: 'mgc_baseball_fill',
-    // 하키: 'mgc_basketball_fill',
-    과잠: 'mgc_coat_fill',
-    자체제작: 'mgc_hand_fill',
-    악세서리: 'mgc_scarf_fill',
-    빈티지: 'mgc_beard_fill',
-    리폼: 'mgc_scissors_2_fill',
-    기타: 'mgc_coathanger_fill',
-    대여: 'mgc_basket_fill',
-    판매: 'mgc_wallet_4_fill',
-    직거래: 'mgc_group_fill',
-    택배거래: 'mgc_package_2_fill',
-  };
-  
-  export default iconMap;
+  축구: 'mgc_football_fill',
+  농구: 'mgc_basketball_fill',
+  야구: 'mgc_baseball_fill',
+  // 하키: 'mgc_basketball_fill',
+  과잠: 'mgc_coat_fill',
+  자체제작: 'mgc_hand_fill',
+  악세서리: 'mgc_scarf_fill',
+  빈티지: 'mgc_beard_fill',
+  리폼: 'mgc_scissors_2_fill',
+  기타: 'mgc_coathanger_fill',
+  대여: 'mgc_basket_fill',
+  판매: 'mgc_wallet_4_fill',
+  직거래: 'mgc_group_fill',
+  택배거래: 'mgc_package_2_fill',
+};
+
+export default iconMap;
