+++ src/pages/PostPages/stepPages/Step2/style.css.ts
@@ -1,8 +1,8 @@
-import { css } from "@styled-system/css";
+import { css } from '@styled-system/css';
 
-export const Container = css ({
-        display: 'grid',
-        gridTemplateColumns: 'repeat(2, 1fr)',
-        padding: '1rem',
-        gap: '0.62rem',
-})
+export const Container = css({
+  display: 'grid',
+  gridTemplateColumns: 'repeat(2, 1fr)',
+  padding: '1rem',
+  gap: '0.62rem',
+});
