+++ src/svg.d.ts
@@ -1,4 +1,4 @@
-declare module "*.svg" {
-    const value: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
-    export default value;
-  }
+declare module '*.svg' {
+  const value: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
+  export default value;
+}
